requirements.txt
# WordPress Publishing Service - Dependencies
# Generated: 2025-10-27

# ==================== Core Dependencies ====================

# Browser Automation
playwright==1.40.0

# AI / Computer Use
anthropic==0.8.1

# Configuration Management
pyyaml==6.0.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# ==================== API Framework ====================

# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client (for testing)
httpx==0.25.2

# ==================== Testing ====================

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-timeout==2.2.0

# Mocking
pytest-mock==3.12.0

# ==================== Utilities ====================

# Image Processing
pillow==10.1.0

# Environment Variables
python-dotenv==1.0.0

# Monitoring & Metrics
prometheus-client==0.19.0

# Date/Time
python-dateutil==2.8.2

# ==================== Development Tools (Optional) ====================

# Code Quality
# black==23.12.1
# flake8==7.0.0
# mypy==1.7.1

# Documentation
# mkdocs==1.5.3
# mkdocs-material==9.5.3
